refactor: controller에 있던 sql query service로 분리
