+++ backend/src/v1/search-keywords/searchKeywords.service.ts
@@ -1,6 +1,8 @@
 import { makeExecuteQuery, pool } from '~/mysql';
 import { logger } from '~/logger';
 import { PopularSearchKeyword, SearchKeyword } from './searchKeywords.type';
+import { extractHangulInitials, disassembleHangul } from '~/v1/utils/disassembleKeywords';
+import { executeQuery } from '~/mysql';
 import * as searchKeywordRepository from './searchKeywords.repository';
 
 const LEAST_SEARCH_COUNT = 5;
@@ -100,3 +102,129 @@ export const createSearchKeywordLog = async (
     connection.release();
   }
 };
+
+
+export const getSearchAutocompletePreviewResult = async (
+  keyword: string
+) => {
+  const LIMIT_OF_SEARCH_keyword_PREVIEW = 12;
+  let keyword_d = extractHangulInitials(keyword as string)
+  let isCho = true;
+
+  if (keyword !== keyword_d) {
+    keyword_d = disassembleHangul(keyword as string);
+    isCho = false;
+  }
+
+  let result: any = [];
+  let totalCount: number;
+  if (isCho) {
+    result = await executeQuery(
+      `
+      (
+        SELECT id as book_info_id, title, author, publisher, publishedAt, image
+        FROM book_info
+        WHERE id IN (
+            SELECT book_info_id
+            FROM book_info_search_keywords
+            WHERE MATCH(title_initials, author_initials, publisher_initials) AGAINST (? IN BOOLEAN MODE)
+        )
+    )
+    UNION (
+        SELECT id as book_info_id, title, author, publisher, publishedAt, image
+        FROM book_info
+        WHERE id IN (
+            SELECT book_info_id
+            FROM book_info_search_keywords
+            WHERE title_initials LIKE ('%${keyword_d}%')
+                    OR author_initials LIKE ('%${keyword_d}%')
+                    OR publisher_initials LIKE ('%${keyword_d}%')
+        )
+    )
+    LIMIT ${LIMIT_OF_SEARCH_keyword_PREVIEW}
+      `,      [keyword_d]
+    );
+    totalCount = await executeQuery(
+      `SELECT COUNT(*) AS totalCount FROM (
+        (
+          SELECT id
+          FROM book_info
+          WHERE id IN (
+              SELECT book_info_id
+              FROM book_info_search_keywords
+              WHERE MATCH(title_initials, author_initials, publisher_initials) AGAINST (? IN BOOLEAN MODE)
+          )
+      )
+      UNION (
+          SELECT id
+          FROM book_info
+          WHERE id IN (
+              SELECT book_info_id
+              FROM book_info_search_keywords
+              WHERE title_initials LIKE ('%${keyword_d}%')
+                      OR author_initials LIKE ('%${keyword_d}%')
+                      OR publisher_initials LIKE ('%${keyword_d}%')
+          )
+      )) AS COUNT_SET`,      [keyword_d]
+      ).then((result) => {
+        return result[0]
+      });
+  } else {
+    result = await executeQuery(
+      
+      `(
+        SELECT id as book_info_id, title, author, publisher, publishedAt, image         
+        FROM book_info
+        WHERE id IN (
+            SELECT book_info_id
+            FROM book_info_search_keywords
+            WHERE MATCH(disassembled_title, disassembled_author, disassembled_publisher) AGAINST (? IN BOOLEAN MODE)
+        )
+    )
+    UNION (
+        SELECT id as book_info_id, title, author, publisher, publishedAt, image         
+        FROM book_info
+        WHERE id IN (
+            SELECT book_info_id
+            FROM book_info_search_keywords
+            WHERE disassembled_title LIKE ('%${keyword_d}%')
+                    OR disassembled_author LIKE ('%${keyword_d}%')
+                    OR disassembled_publisher LIKE ('%${keyword_d}%')
+        )
+    )
+    LIMIT ${LIMIT_OF_SEARCH_keyword_PREVIEW}
+      `,      [keyword_d]
+    );
+    totalCount = await executeQuery(
+      `SELECT COUNT(*) AS totalCount FROM (
+        (
+          SELECT id
+          FROM book_info
+          WHERE id IN (
+              SELECT book_info_id
+              FROM book_info_search_keywords
+              WHERE MATCH(disassembled_title, disassembled_author, disassembled_publisher) AGAINST (? IN BOOLEAN MODE)
+          )
+      )
+      UNION (
+          SELECT id
+          FROM book_info
+          WHERE id IN (
+              SELECT book_info_id
+              FROM book_info_search_keywords
+              WHERE disassembled_title LIKE ('%${keyword_d}%')
+                      OR disassembled_author LIKE ('%${keyword_d}%')
+                      OR disassembled_publisher LIKE ('%${keyword_d}%')
+          )
+      )) AS COUNT_SET`,      [keyword_d]
+    ).then((result) => {
+      return result[0]
+    });
+    
+  }
+  return {
+    items: result,
+    meta: 
+      totalCount
+  };
+}
